refactor:typeとutils用のディレクトリを作成
