+++ src/test/utils/convertHyphenatedZipCode.test.ts
@@ -1,8 +1,8 @@
 import { expect } from '@jest/globals';
-import { goqZipCode } from './goqZipCode';
-import { addresses } from './constants/address';
-import { addressExcludingHyphenDataList } from './constants';
-import { addressIncludingHyphenDataList } from './constants';
+import { goqZipCode } from '../goqZipCode';
+import { addresses } from '../constants/address';
+import { addressExcludingHyphenDataList } from '../constants';
+import { addressIncludingHyphenDataList } from '../constants';
 
 describe('オプションによってハイフンを付与', () => {
   test('オプションでハイフンありを指定している場合、郵便番号にハイフンを追加する', () => {
